:recycle: refactor : LMS 디렉토리 하위로 이동
