+++ src/main/java/nextstep/lms/domain/Image.java
@@ -1,4 +1,4 @@
-package nextstep.images.domain;
+package nextstep.lms.domain;
 
 import java.util.Objects;
