+++ src/test/java/nextstep/courses/domain/CourseTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.lms.domain.Course;
 import nextstep.sessions.domain.SessionTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
