+++ src/test/java/nextstep/enrollment/domain/EnrollmentTest.java
@@ -1,5 +1,6 @@
 package nextstep.enrollment.domain;
 
+import nextstep.lms.domain.Enrollment;
 import nextstep.sessions.domain.SessionTest;
 import nextstep.users.domain.NsUserTest;
 import org.junit.jupiter.api.DisplayName;
